Revert "[refactor] Content, Timer, UrlInput 별로 Layout 분리"
This reverts commit 4ca808b76edc51bb5d37df14e1f33b4c7afb9a56.
