+++ src/layouts/UrlInputLayout.tsx
@@ -1,0 +1,12 @@
+import styled, { css } from "styled-components";
+
+const InputLayout = styled.input`
+  border-radius: 4px;
+  border: 1px solid #dee2e6;
+  width: 100%;
+  outline: none;
+  font-size: 18px;
+  box-sizing: border-box;
+`;
+
+export { InputLayout };
